chore: extract channel
